refactor(kernel): promove ReferenciaCidadeGeo para o Kernel compartilhado
A validação da referência de cidade do Geo (ReferenciaCidadeGeo +
CidadeReferenciaErrorCodes) vivia em Configuracao.Domain. Para que a
Instituicao (módulo organizacao-institucional) adote o mesmo padrão de
referência por composição no cliente (ADR-0090), ela precisa do validador
sem depender de Configuracao — o isolamento cross-módulo (R8 / ADR-0056)
proíbe esse using.

Promove ambos para Unifesspa.UniPlus.Kernel (namespace Kernel.Domain.Cidades),
ao lado dos value objects de domínio compartilhados já existentes (Cpf,
Email, NomeSocial). Campus e LocalOferta re-apontam os usings; o teste
migra para Kernel.UnitTests. Refactor puro, sem mudança de comportamento.

Em service de #686.
