+++ src/configuracao/Unifesspa.UniPlus.Configuracao.Application/Commands/Campi/AtualizarCampusCommandValidator.cs
@@ -2,7 +2,7 @@ namespace Unifesspa.UniPlus.Configuracao.Application.Commands.Campi;
 
 using FluentValidation;
 
-using Unifesspa.UniPlus.Configuracao.Domain.Cidades;
+using Unifesspa.UniPlus.Kernel.Domain.Cidades;
 
 public sealed class AtualizarCampusCommandValidator : AbstractValidator<AtualizarCampusCommand>
 {
