+++ src/configuracao/Unifesspa.UniPlus.Configuracao.Application/Commands/LocaisOferta/CriarLocalOfertaCommandHandler.cs
@@ -1,7 +1,7 @@
 namespace Unifesspa.UniPlus.Configuracao.Application.Commands.LocaisOferta;
 
 using Unifesspa.UniPlus.Application.Abstractions.Interfaces;
-using Unifesspa.UniPlus.Configuracao.Domain.Cidades;
+using Unifesspa.UniPlus.Kernel.Domain.Cidades;
 using Unifesspa.UniPlus.Configuracao.Domain.Entities;
 using Unifesspa.UniPlus.Configuracao.Domain.Errors;
 using Unifesspa.UniPlus.Configuracao.Domain.Interfaces;
