+++ tests/Unifesspa.UniPlus.Configuracao.Application.UnitTests/Commands/CriarCampusCommandHandlerTests.cs
@@ -9,6 +9,7 @@ using Unifesspa.UniPlus.Configuracao.Application.Commands.Campi;
 using Unifesspa.UniPlus.Configuracao.Domain.Entities;
 using Unifesspa.UniPlus.Configuracao.Domain.Errors;
 using Unifesspa.UniPlus.Configuracao.Domain.Interfaces;
+using Unifesspa.UniPlus.Kernel.Domain.Cidades;
 using Unifesspa.UniPlus.Kernel.Results;
 
 public sealed class CriarCampusCommandHandlerTests
